+++ frontend/src/components/Input/Input.css
@@ -1,16 +1,5 @@
-.field {
-  width: 100%;
-  height: 56px;
-  border-radius: 4px;
-  position: relative;
-  background-color: rgba(255,255,255,0.3);
-  transition: 0.3s all;
-}
-.field:hover {
-  background-color: rgba(255, 255, 255, 0.45);
-  box-shadow: 0px 4px 20px 0px rgba(0, 0, 0, 0.05);
-}
 
+/* Tabs Styling */
 .nav-tabs {
     border-bottom: 4px solid #444ACA;
 }
@@ -37,4 +26,19 @@
 	color: white !important;
 	background-color: #C5C5C5!important;
 	font-weight: bold;
+}
+
+/* Textarea Styling */
+textarea.field {
+  width: 100%;
+  resize: none;
+  height: 50vh; /*Update height to reflect grandpare div size*/
+  border: none !important;
+  border:1px solid red;
+}
+
+
+textarea.field:focus {
+    outline: none !important;
+    border:1px solid red;
 }
